// express/schemas/schema.d.ts
export interface components {
  schemas: {
// ...
    GetEvent: {
// ...
      images: {
        idUpload: components['schemas']['Id']
        url: components['schemas']['Url']
      }[]
      artists: components['schemas']['ListArtists']
    }
// ...
    UpdateUserRole: {
      role: components['schemas']['UserRole']
    }
    UpdateUser: {
      name: components['schemas']['Name']
      surname: components['schemas']['Name']
      birthdate: components['schemas']['Date']
    }
    ListUsers: {
      idUser: components['schemas']['Id']
      publicAddress: components['schemas']['PublicAddress']
      email: components['schemas']['Email']
      username: components['schemas']['Name']
      role: components['schemas']['UserRole']
      active: boolean
      created: components['schemas']['DateTime']
    }[]
// ...
    CreateReview: {
      title: components['schemas']['Name']
      eventLocation: components['schemas']['LongName']
      eventDate: components['schemas']['Date']
      content: components['schemas']['Content']
      rate: components['schemas']['Rate']
    }
// ...
  }
  responses: {
    /** @description Invalid request contents */
    BadRequest: {
      content: {
        'application/json': components['schemas']['Error']
      }
    }
// ...
  }
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}

export type $defs = Record<string, never>

export type external = Record<string, never>

export interface operations {
// ...
  ListUsers: {
    parameters: {
      query?: {
        /** @description A string to search user by username */
        search?: string
        /** @description A listing page number */
        page?: components['schemas']['Page']
      }
    }
    responses: {
      /** @description Available users */
      200: {
        content: {
          'application/json': {
            count: number
            pages: number
            currentPage: number
            results: components['schemas']['ListUsers']
          }
        }
      }
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Generate nonce for user with provided wallet address */
  GenerateNonce: {
    requestBody?: {
      content: {
        'application/json': {
          publicAddress: components['schemas']['PublicAddress']
        }
      }
    }
    responses: {
      /** @description User nonce */
      200: {
        content: {
          'application/json': {
            nonce: components['schemas']['Nonce']
          }
        }
      }
      400: components['responses']['BadRequest']
      /** @description User not found, return sign up token */
      404: {
        content: {
          'application/json': {
            signUpToken: components['schemas']['JWT']
          }
        }
      }
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Sign up user account */
  SignUp: {
    requestBody?: {
      content: {
        'application/json': {
          signUpToken: components['schemas']['JWT']
          email: components['schemas']['Email']
          username: components['schemas']['Name']
          name: components['schemas']['Name']
          surname: components['schemas']['Name']
          birthdate: components['schemas']['Date']
        }
      }
    }
    responses: {
      /** @description Successfully signed up user account */
      204: {
        content: never
      }
      400: components['responses']['BadRequest']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Activate user account */
  ActivateUser: {
    requestBody?: {
      content: {
        'application/json': {
          activationToken: components['schemas']['JWT']
        }
      }
    }
    responses: {
      /** @description Successfully activated user account */
      204: {
        content: never
      }
      400: components['responses']['BadRequest']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Sign in user */
  SignIn: {
    requestBody?: {
      content: {
        'application/json': {
          message: string
          /** @description ECDSA private key signature of message */
          signature: string
        }
      }
    }
    responses: {
      /** @description JWT tokens pair */
      200: {
        content: {
          'application/json': {
            authenticationToken: components['schemas']['JWT']
            refreshToken: components['schemas']['JWT']
          }
        }
      }
      400: components['responses']['BadRequest']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Refresh JWT tokens pair */
  RefreshTokens: {
    requestBody?: {
      content: {
        'application/json': {
          refreshToken: components['schemas']['JWT']
        }
      }
    }
    responses: {
      /** @description JWT tokens pair */
      200: {
        content: {
          'application/json': {
            authenticationToken: components['schemas']['JWT']
            refreshToken: components['schemas']['JWT']
          }
        }
      }
      400: components['responses']['BadRequest']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Update user role */
  UpdateUserRole: {
    parameters: {
      path: {
        /** @description The unique identifier of the user */
        id: components['schemas']['Id']
      }
    }
    requestBody?: {
      content: {
        'application/json': components['schemas']['UpdateUserRole']
      }
    }
    responses: {
      /** @description Updated user details */
      200: {
        content: {
          'application/json': components['schemas']['GetUser']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description List events created by currently signed in user */
  ListUserEvents: {
    parameters: {
      query?: {
        /** @description A string to search event by name, tags and description */
        search?: string
        /** @description The unique identifier of the subcategory */
        idSubcategory?: components['schemas']['Id']
        /** @description The unique identifier of the city */
        idCity?: components['schemas']['Id']
        /** @description A listing page number */
        page?: components['schemas']['Page']
      }
    }
    responses: {
      /** @description List of events */
      200: {
        content: {
          'application/json': {
            count: number
            pages: number
            currentPage: number
            results: components['schemas']['ListEvents']
          }
        }
      }
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
// ...
  UpdateUser: {
    requestBody?: {
      content: {
        'application/json': components['schemas']['UpdateUser']
      }
    }
    responses: {
      /** @description User updated details */
      200: {
        content: {
          'application/json': components['schemas']['GetUser']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      default: components['responses']['InternalServerError']
    }
  }
// ...
}